fix: grpc fallback to websocket
